Renamed middleware namespace and classes
